define a package and a writer for each format
