+++ report/html/writer.go
@@ -1,0 +1,17 @@
+package html
+
+import (
+	"github.com/securego/gosec/v2/report/core"
+	"html/template"
+	"io"
+)
+
+//WriteReport write a report in html format to the output writer
+func WriteReport(w io.Writer, data *core.ReportInfo) error {
+	t, e := template.New("gosec").Parse(Template)
+	if e != nil {
+		return e
+	}
+
+	return t.Execute(w, data)
+}
