+++ report/yaml/writer.go
@@ -1,0 +1,17 @@
+package yaml
+
+import (
+	"github.com/securego/gosec/v2/report/core"
+	"gopkg.in/yaml.v2"
+	"io"
+)
+
+//WriteReport write a report in yaml format to the output writer
+func WriteReport(w io.Writer, data *core.ReportInfo) error {
+	raw, err := yaml.Marshal(data)
+	if err != nil {
+		return err
+	}
+	_, err = w.Write(raw)
+	return err
+}
